Math controller Refactored (without dependency injection) and added new operations
