--- rest-api-with-spring-boot/src/main/java/com/gabriel/MathController.java
@@ -1,40 +1,0 @@
-package com.gabriel;
-
-import com.gabriel.exceptions.UnsupportedMathOperationException;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.concurrent.atomic.AtomicLong;
-
-
-@RestController
-public class MathController {
-
-    private AtomicLong counter = new AtomicLong();
-
-    @RequestMapping(value = "/sum/{numberOne}/{numberTwo}", method = RequestMethod.GET)
-    public double sum(
-            @PathVariable(value = "numberOne") String numberOne,
-            @PathVariable(value = "numberTwo") String numberTwo
-    ) throws Exception{
-
-        if (!isNumeric(numberOne) || !isNumeric(numberTwo)) {
-            throw new UnsupportedMathOperationException("Please set a numeric value!");
-        }
-        return convertDouble(numberOne) + convertDouble(numberTwo);
-    }
-
-    private double convertDouble(String strNumber) {
-        if (strNumber == null) return 0D;
-
-        String number = strNumber.replaceAll(",", ".");
-        if (isNumeric(number)) return Double.parseDouble(number);
-        return 0D;
-    }
-
-
-    private boolean isNumeric(String strNumber) {
-        if (strNumber == null) return false;
-        String number = strNumber.replaceAll(",", ".");
-        return number.matches("[-+]?[0-9]*\\.?[0-9]+");
-    }
-}
