+++ rest-api-with-spring-boot/src/main/java/com/gabriel/controllers/MathController.java
@@ -1,0 +1,69 @@
+package com.gabriel.controllers;
+
+import com.gabriel.exceptions.UnsupportedMathOperationException;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.concurrent.atomic.AtomicLong;
+
+import static com.gabriel.converter.MathParamsConverter.convertDouble;
+import static com.gabriel.converter.MathParamsConverter.isNumeric;
+
+
+@RestController
+public class MathController {
+
+    private AtomicLong counter = new AtomicLong();
+
+    //URL e método de soma
+    @RequestMapping(value = "/sum/{numberOne}/{numberTwo}", method = RequestMethod.GET)
+    public double sum(
+            @PathVariable(value = "numberOne") String numberOne,
+            @PathVariable(value = "numberTwo") String numberTwo
+    ) throws Exception{
+
+        if (!isNumeric(numberOne) || !isNumeric(numberTwo)) {
+            throw new UnsupportedMathOperationException("Please set a numeric value!");
+        }
+        return convertDouble(numberOne) + convertDouble(numberTwo);
+    }
+
+    //URL e método de subtração
+    @RequestMapping(value = "/subtraction/{numberOne}/{numberTwo}", method = RequestMethod.GET)
+    public double subtraction(
+            @PathVariable(value = "numberOne") String numberOne,
+            @PathVariable(value = "numberTwo") String numberTwo
+    ) throws Exception{
+
+        if (!isNumeric(numberOne) || !isNumeric(numberTwo)) {
+            throw new UnsupportedMathOperationException("Please set a numeric value!");
+        }
+        return convertDouble(numberOne) - convertDouble(numberTwo);
+    }
+
+    //URL e método de multiplicação
+    @RequestMapping(value = "/multiplication/{numberOne}/{numberTwo}", method = RequestMethod.GET)
+    public double multiplication(
+            @PathVariable(value = "numberOne") String numberOne,
+            @PathVariable(value = "numberTwo") String numberTwo
+    ) throws Exception{
+
+        if (!isNumeric(numberOne) || !isNumeric(numberTwo)) {
+            throw new UnsupportedMathOperationException("Please set a numeric value!");
+        }
+        return convertDouble(numberOne) * convertDouble(numberTwo);
+    }
+
+    //URL e método de divisão
+    @RequestMapping(value = "/division/{numberOne}/{numberTwo}", method = RequestMethod.GET)
+    public double division(
+            @PathVariable(value = "numberOne") String numberOne,
+            @PathVariable(value = "numberTwo") String numberTwo
+    ) throws Exception{
+
+        if (!isNumeric(numberOne) || !isNumeric(numberTwo)) {
+            throw new UnsupportedMathOperationException("Please set a numeric value!");
+        }
+        return convertDouble(numberOne) / convertDouble(numberTwo);
+    }
+
+}
