+++ rest-api-with-spring-boot/src/main/java/com/gabriel/converter/MathParamsConverter.java
@@ -1,0 +1,20 @@
+package com.gabriel.converter;
+
+public class MathParamsConverter {
+
+    public static double convertDouble(String strNumber) {
+        if (strNumber == null) return 0D;
+
+        String number = strNumber.replaceAll(",", ".");
+        if (isNumeric(number)) return Double.parseDouble(number);
+        return 0D;
+    }
+
+
+    public static boolean isNumeric(String strNumber) {
+        if (strNumber == null) return false;
+        String number = strNumber.replaceAll(",", ".");
+        return number.matches("[-+]?[0-9]*\\.?[0-9]+");
+    }
+
+}
